feat(lib): prompt to run git config commands on startup
So it's a little annoying to have to reach for the mouse to copy and paste
the suggested git config commands so let's just suggest them and also prompt
to run them with the press of a button. The "auto" guessing code for the
values is working pretty well so let's do it.